+++ test/fixtures/assoc.rb
@@ -14,3 +14,27 @@
 }
 -
 { foo: bar }
+%
+{
+  foo: [
+    fooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+  ]
+}
+%
+{
+  foo: {
+    fooooooooooooooooooooooooooooooooo: ooooooooooooooooooooooooooooooooooooooo
+  }
+}
+%
+{
+  foo: -> do
+    foooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+  end
+}
+%
+{
+  foo: %w[
+    foooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+  ]
+}